Uploaded Astropy tutorial
Here is the Astropy Tutorial. When running the tutorial make sure to change the path of the files on your laptop. Right now, it is set as my desktop path "eyune/Desktop", yours should be "name/Desktop" or whichever path you choose.

I also uploaded the images you should obtain.

Note: Sometimes when I chose to Restart and Run All it wouldn't output the actual figure size. To fix this, just re-run manually from the first cell where we define the figure size.